Add README and fix jar ignore
