--- .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Tue Jun 10 03:12:28 MSK 2025
-gradle.version=8.6
