--- src/main/java/Main.java
@@ -1,12 +1,0 @@
-import org.telegram.telegrambots.meta.TelegramBotsApi;
-import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
-
-public class Main {
-    public static void main(String[] args) throws Exception {
-        TelegramBotsApi botsApi = new TelegramBotsApi(DefaultBotSession.class);
-        botsApi.registerBot(new TelegramVacancyBot());
-    }
-}
-
-
-
