+++ src/main/java/org/example/autobot/Main.java
@@ -1,0 +1,9 @@
+package org.example.autobot;
+import org.telegram.telegrambots.meta.TelegramBotsApi;
+import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
+public class Main {
+    public static void main(String[] args) throws Exception {
+        TelegramBotsApi botsApi = new TelegramBotsApi(DefaultBotSession.class);
+        botsApi.registerBot(new TelegramVacancyBot());
+    }
+}
